Rounded total pages up with Math.ceil
I had a problem with not being allowed to see the last page of rows. I had 14 total rows, showing 10 per page. When doing the math, you get 1.4 and I suppose the original code was saying that was one page. I added the javascript Math.ceil function to round up to 2 pages and it solved my problem.